refactor(session): extract prompt parts
